Rename CodeEditor to SourceCode
